Question Order

+++ src/test/java/jorge/rv/QuizZz/unitTests/service/usermanagement/utils/DateHelperTests.java
@@ -13,8 +13,8 @@ import jorge.rv.quizzz.service.usermanagement.utils.DateHelperImpl;
 public class DateHelperTests {
 
 	private static Date FIXED_DATE = new Date(123456);
-	private static int VALID_OFFSET = 5;
-	private static int NEGATIVE_OFFSET = -5;
+	private static long VALID_OFFSET = 5;
+	private static long NEGATIVE_OFFSET = -5;
 	
 	private DateHelper dateHelper;
 	
@@ -25,14 +25,14 @@ public class DateHelperTests {
 	
 	@Test
 	public void calculateValidExpirationDate() {
-		Date expirationDate = dateHelper.getExpirationDate(FIXED_DATE, VALID_OFFSET);
+		Date expirationDate = dateHelper.getExpirationDate(FIXED_DATE, (int) VALID_OFFSET);
 		
 		assertEquals(VALID_OFFSET * 1000 * 60, expirationDate.getTime() - FIXED_DATE.getTime());
 	}
 	
 	@Test
 	public void calculateNegativeExpirationDate_shouldReturnFromDate() {
-		Date expirationDate = dateHelper.getExpirationDate(FIXED_DATE, NEGATIVE_OFFSET);
+		Date expirationDate = dateHelper.getExpirationDate(FIXED_DATE, (int) NEGATIVE_OFFSET);
 		
 		assertEquals(0, expirationDate.getTime() - FIXED_DATE.getTime());
 	}
